renamed:    src/components/navbar-about.vue -> src/components/navbar-application.vue 	modified:   src/components/navbar.vue 	modified:   src/components/quoteform.vue 	deleted:    src/components/socials.vue 	modified:   src/router/index.js 	deleted:    src/views/AboutView.vue 	new file:   src/views/ApplicationView.vue 	modified:   src/views/HomeView.vue
