--- src/components/socials.vue
@@ -1,68 +1,0 @@
-<template>
-  <div class="socials">
-    <div class="socials-2">
-      <img
-        loading="lazy"
-        src="https://cdn.builder.io/api/v1/image/assets/TEMP/3686df59b825581b081ab7040e0ce58e3aaf0c63ffdec76bec0d76e8c304ded4?"
-        class="img"
-      /><img
-        loading="lazy"
-        src="https://cdn.builder.io/api/v1/image/assets/TEMP/c944afb5807c91d1764551c36a3a4857f3abcd4286ca8fd84f1788c9b03882d7?"
-        class="img-2"
-      /><img
-        loading="lazy"
-        src="https://cdn.builder.io/api/v1/image/assets/TEMP/1f46e4119be644b8fa17b82f06967ce05add8793c03c342d968b38b0e27e019f?"
-        class="img"
-      /><img
-        loading="lazy"
-        src="https://cdn.builder.io/api/v1/image/assets/TEMP/49ff79a85c5d1a6413352b6a3a3f13fb511b47d53f571051003c2da0ad330b7d?"
-        class="img"
-      /><img
-        loading="lazy"
-        src="https://cdn.builder.io/api/v1/image/assets/TEMP/7c346767b750fa13e383d21c98dfae9330cf42ea428365d8489eaec80d0560ae?"
-        class="img-3"
-      /><img
-        loading="lazy"
-        src="https://cdn.builder.io/api/v1/image/assets/TEMP/8f63b9c81d58f60f7a8660bf3b66b00f3cef5893bf1f7aafca573687fcdab4f8?"
-        class="img"
-      />
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.socials {
-  justify-content: center;
-  align-items: center;
-  background-color: #1e1e1e;
-  display: flex;
-  padding: 28px 60px;
-}
-
-.socials-2 {
-  display: flex;
-  width: 640px;
-  max-width: 100%;
-  gap: 20px;
-  justify-content: space-between;
-  cursor: pointer;
-}
-.img {
-  aspect-ratio: 1;
-  object-fit: auto;
-  object-position: center;
-  width: 22px;
-}
-.img-2 {
-  aspect-ratio: 0.95;
-  object-fit: auto;
-  object-position: center;
-  width: 21px;
-}
-.img-3 {
-  aspect-ratio: 0.95;
-  object-fit: auto;
-  object-position: center;
-  width: 21px;
-}
-</style>
